// swiyu-ecosystem-portal-ui/src/app/api/generated/model/business-partner.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Address} from './address';
import {BusinessPartnerTrustStatus} from './business-partner-trust-status';

export interface BusinessPartner {
  /**
   * Key of the partner under which it is registered with ePortal
   */
  id: string;
  /**
   * Unique name of the partner
   */
  name: string;
  contactEmailAddress: string;
  /**
   * Type of the partner
   */
  type: BusinessPartner.TypeEnum;
  /**
   * User paid for trust onboarding
   */
  payedForTrustVerification?: boolean;
  /**
   * Number of DID slots the user paid for already
   */
  payedForDIDSlots?: number;
  createdAt?: string;
  updatedAt?: string;
  /**
   * Enterprise identification number of the partner
   */
  uid?: string;
  /**
   * Address of the partner
   */
  address?: Address;
  contactPhone: string;
  /**
   * Aggregated state of the trust process for this business partner
   */
  trustVerificationStatus?: BusinessPartnerTrustStatus;
  /**
   * Time limit, if necessary, of the current aggregated state of the trust process
   */
  maxDateForTrustVerificationStatus?: string;
}
export namespace BusinessPartner {
  export const TypeEnum = {
    GovernmentalInstitution: 'GOVERNMENTAL_INSTITUTION',
    Business: 'BUSINESS',
    Individual: 'INDIVIDUAL',
    Unknown: 'UNKNOWN'
  } as const;
  export type TypeEnum = (typeof TypeEnum)[keyof typeof TypeEnum];
}